@@ -1,13 +1,12 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
-
   <PropertyGroup>
-    <TargetFramework>netcoreapp2.2</TargetFramework>
-    <AspNetCoreHostingModel>InProcess</AspNetCoreHostingModel>
+    <TargetFramework>netcoreapp3.1</TargetFramework>
   </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.AspNetCore.App" />
-    <PackageReference Include="Microsoft.AspNetCore.Razor.Design" Version="2.2.0" PrivateAssets="All" />
+  <ItemGroup>&#xD;
+&#xD;
   </ItemGroup>
-
+  <ItemGroup>
+    <PackageReference Include="System.Net.Http.Json" Version="3.2.1"/>
+    <PackageReference Include="LettuceEncrypt" Version="1.1.0-beta.73"/>
+  </ItemGroup>
 </Project>
